refactor: strongly type governance field on EventMetadata
Extract SessionMeta and related result types into governance/results.py
to break the circular import between types.py and pipeline.py. The
governance field on EventMetadata is now SessionMeta | None (not Any or
dict). All sinks use typed attribute access (gov.recommendation.
recommended_action.value) instead of dict.get() chains.
